Refactor svg_mobject.py into multiple files.
